create advanced results middlware for use on all resources
